pyproject.toml
[project]
name = "isa4rl-article"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "datetime>=6.0",
    "gymnasium>=1.2.3",
    "highway-env>=1.10.2",
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "pandas>=3.0.1",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.0",
    "seaborn>=0.13.2",
    "stable-baselines3>=2.7.1",
    "tqdm>=4.67.3",
]
